Bunch of cleanup type work.  Decided to package together all required icons regardless of if they are available elsewhere.  Added error page, apparantly I do indeed need one too.  Basically some adapting work to IntelliJ 6.0.
git-svn-id: http://svn.igniterealtime.org/svn/repos/wildfire/trunk/src/plugins/gateway@5653 b35dd754-fafc-0310-a699-88a17e54d16e

+++ src/java/org/jivesoftware/wildfire/gateway/protocols/oscar/SnacManager.java
@@ -49,7 +49,7 @@ public class SnacManager {
 
             if (handlers == null) {
                 handlers = new LinkedList<BasicFlapConnection>();
-                conns.put((Integer)familyCode, handlers);
+                conns.put(familyCode, handlers);
             }
 
             if (!handlers.contains(conn)) {
